Feature/55 home server calender

+++ app/src/main/java/kr/co/nottodo/view/calendar/weekly/WeeklyCalendarView.kt
@@ -137,7 +137,7 @@ class WeeklyCalendarView @JvmOverloads constructor(
     }
 
     /** 서버 통신된 NotToDo 갯수 갱신 **/
-    fun setNotToDoCount(list: List<Pair<LocalDate?, Double>>) {
+    fun setNotToDoCount(list: List<Pair<LocalDate?, Float>>) {
         weeklyCalendar.setNotToDoCount(
             list = list
         )
